pyproject.toml
[tool.poetry]
name = "llmsat"
version = "0.1.0"
description = "An LLM-based autonomous agent for space exploration"
authors = ["David Maranto <D.Maranto1122@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.278"
openai = "^0.27.10"
python-decouple = "^3.8"
pydantic = "^2.3.0"
pytest = "^7.4.1"
astropy = "^5.3.2"
krpc = "^0.5.3"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
